Context FS stuff (#47)
* Context FS stuff

* Fleshing it out

* Discard tests

* More mocks

* GoLand crap

* Adapter specs

* Setter fs tests

* AccessorFs tests

* Refactored variable naming in AccessorFs

The variable 'fs' in the AccessorFs struct has been renamed to 'Fs'. This change is reflected across all methods within the struct where this variable was used. The purpose of this refactoring is to improve code readability and maintain consistency with Go's naming conventions.

* Refactor filesystem interface methods

The filesystem interface methods have been refactored to include 'Context' in their names for better clarity. The comments for these methods have also been updated accordingly. Additionally, the redundant comment block for the Fs interface has been removed and the AferoFs interface has been renamed to align with its purpose.

* Refactor the docker fs