Add a version command to the devops CLI (#59)
* Fail

* Little bit of nonsense in my life

* Versions

* Working and versioning and paths and contexts

* Refactor Context struct to Directory type

The 'Context' struct has been refactored into a simpler 'Directory' type. This change simplifies the code by removing unnecessary methods and properties, and improves readability. The functions Git(), Cwd(), and Load() have been updated to return a Directory instead of a pointer to a Context object. Corresponding changes have also been made in the test file.

* Refactor version reading and error handling

Simplified the file reading process in the version package by removing unnecessary options. The 'root' option has been removed, and the workspace context is no longer used. Error handling for missing dependencies has also been improved. Tests have been updated to reflect these changes, with a focus on ensuring correct behavior when a dependency does not exist.

* Added test suite and enhanced version command

Introduced a new test suite for the cmd package. The version command has been improved to include VersionOptions struct, which allows changing the working directory before execution. Also added tests for these new options in the version command.

* Added directory change options to devops tests

The test suite for the devops command now includes checks for both long and short forms of the option to change directories. This ensures that the command behaves as expected when asked to operate in a different directory.