Refactor ChdirOptions and add new functions
The ChdirOptions struct has been refactored to improve its functionality. The 'Chdir' field was renamed to 'dir'. A new function, 'Chdir', was added to change the current working directory. Another function, 'NewChdirOptions', was introduced for creating a new instance of ChdirOptions. These changes also required modifications in the associated tests.