Init versions (#60)
* Updated version command and flag handling

The 'version' command now accepts an optional name argument. Additionally, the 'chdir' flag has been marked to accept directory names. This improves usability by allowing users to specify a dependency for version checking and change directories before execution.

* Added directory change functionality and tests

Introduced a new feature that allows changing to a specified directory before executing commands. This is achieved through the addition of ChdirOptions struct and associated methods in cmd.go. Also, added corresponding unit tests to ensure the correct working of this new feature.

* Removed redundant test files and simplified version command

Significant changes include:
- Deleted two unnecessary test files
- Simplified the VersionOptions struct in the version command by removing a function and replacing it with an existing one from another package
- Removed a flag related to directory change from the NewVersion function, as it was no longer needed after simplifying VersionOptions.

* Optimized lazy evaluation functions

Removed the Lift function from the lazy evaluation module as it was redundant. Introduced a new Value function to extract values from Lazy types. Also, cleaned up some code and added comments for better readability.

* Refactor ChdirOptions and add new functions

The ChdirOptions struct has been refactored to improve its functionality. The 'Chdir' field was renamed to 'dir'. A new function, 'Chdir', was added to change the current working directory. Another function, 'NewChdirOptions', was introduced for creating a new instance of ChdirOptions. These changes also required modifications in the associated tests.

* Added version management functionality

Introduced a new package for managing versions. This includes the ability to initialize a version directory, read from it, and determine the latest version from a source. Also added support for guessing the source of the version based on input string. Currently, only GitHub is recognized as a valid source.

* Refactor version command error handling

Simplified the error handling in the version command. Removed unnecessary os package import and replaced it with a direct call to opts.Chdir(). This change reduces complexity and improves readability of the code.

* Added versioning functionality

A new feature has been added to handle versioning of dependencies. This includes the ability to generate files for specific dependency versions. Two sources for version information have been defined: 'auto' and 'github'. The 'auto' source guesses the version, while the 'github' source retrieves it directly from GitHub. Error handling has also been improved in case of unsupported sources or other issues during initialization.

* Added 'init' command to devops tool

The new 'init' command has been added to the devops CLI tool. This command is designed to generate files for a specified scaffold. The implementation includes creating a new file for the 'init' command and adding it to the main function of the devops tool.

* Some of the cleanest goddamn code I've ever written

* Testing test test

* WIP on e2e tests

* Added Git testing utilities

Introduced two new functions in the testing package: `GitInit` and `TempGitRepo`. The former initializes a git repository at a given path, while the latter creates a temporary directory and initializes it as a git repository. Also added an interface 'T' with method TempDir() for creating temporary directories.

* gomodtidy

* Refactored git initialization in tests

The code changes refactor the way git repositories are initialized in test cases. Instead of manually executing a 'git init' command for each test, a helper function from the testing package is now used. This simplifies the setup process and makes the tests more readable and maintainable.

* Refactored version handling and added tests

- Moved Regex variable from source.go to version.go
- Added Clean function in version.go to trim leading 'v's from versions
- Relocated RelPath function within the file for better organization
- Created new test file version_test.go to validate Regex, RelPath, and Clean functions

* Updated version initialization

The version initialization function has been updated to clean the version string before writing it to a file. This change ensures that any unwanted characters are removed from the version string, improving data consistency and reliability.

* Enhanced version initialization in tests

Expanded the test coverage for version initialization. Added new tests to handle inline versions, prefix stripping, source specification and error handling. Also modified the command line arguments parsing logic in 'version.go' to accommodate these changes.