Tokenize more like a tokenizer (#65)
* Buncha definitions and some tokens tests

* Refactoring scan, potentially for the worse

* Break everything

* Parsing progress

* Yet another split function

* I present, the new parser

* Whole lotta parsing

* More scanners

* More WIP on new scanner

* Refactor parsing logic and update tests

The parsing logic has been moved from make.go to a new file, parse.go. The 'Parse' function now logs unexpected tokens. The test file has been renamed to match the new structure and updated accordingly.

In scan.go, the Scanner struct is marked as deprecated with a note that it's still a work in progress. The Scan method has been simplified by removing commented-out code and restructuring the remaining code for clarity.

An unused import in make.go was also removed.

* Clean up and implement it in a dumb way